Create status response class (.hpp and .cpp files)
With the creation of this header and source file, the Response class for
the API "status" operation (a.k.a progress tracking) has been declared
and defined. Now, all the Response operations are split in different
files, compared to the previous single "TapeResponse" class, handling
all the API operations together.