Create release response class (.hpp and .cpp)
With the creation of this header and source file, the Response class for
the API "Release" operation has been declared and defined. Now, all the
Response operations are split in different files, compared to the
previous single "TapeResponse" class, handling all the API operations
together.
This API operation is still under work in progress, and will probably be
deprecated due to storm impossiblity to release staged files from cache.