Implement file locality properly
The implementation relies exclusively on:
- stubbification: the size of a file is more than the space occupied
  by the blocks allocated for the file
- presence of a storm.migrated xattr
Hence there is no need to rely on GPFS-specific libraries.

About stubbification: the check could be simplified relying on the
number of blocks being zero. The current check may mean that the file
is being recalled but the recall has not completed yet or it failed.