Distinguish between logical and physical paths
A logical path is a path passed by a client, e.g. in a stage request
or in an archiveinfo request. A logical path is subject to resolution
via the Storage Areas mapping and, if successful, produces a physical
path, which is used for all the actual operations on the file system,
including being passed to GEMSS.

The physical path is also kept in the database, which has required a
change in the schema in order to add an additional column. Note that
in this way the translation happens when the stage request is received
and not subsequently, even if the Storage Areas mapping changes.

Moreover:
- Adjust error messages to make them more in line with the equivalent
  for errno
- Return 500 - internal error in case there is failure to include a
  request in the database. It has happened for example following the
  change in the schema