Augment versioning with git describe
Extend versioning information with the output of git describe. This
additional information includes the number of commits after an
annotated tag (which we assume corresponds to a versioning tag on the
version.txt file), a commit hash and a "dirty" label. This extended
information should be empty when pushing a versioning tag. The "dirty"
label should be present only in a developer's workspace.

The above will be useful when versioning packages, especially in a CI
context.

Keep the base version in a text file rather than in cmake syntax.