Add support for RPM packaging
The RPM contains both the release and debug versions of the storm-tape
executable. This requires the use of a multi-config generator for
cmake.

The corresponding job in CI is triggered only by changes in the
sources and in the version information. To trigger an additional
build, change the release number in CMakeLists.txt.