Consider also the port in the Forwarded header
Even if it refers to the storm-tape service running in the back-end and
not to NGINX, the default port is taken from the configuration, where it
is currently hardcoded to 8080. The port can be changed via the
Forwarded header, either setting it explicitly or being deduced from the
scheme (proto): https deduces 443, http deduces 80.

Also:
- Rename get_host to get_hostinfo
- Restructure implementation
- Add a few tests for the processing of the Forwarded header

Resolves: #3