Include the user (principal) in the access log
The principal is extracted from a request header, either as the `sub`
claim of a JWT access token (header `x-sub`) or the Subject DN
corresponding to the VOMS user attribute of a proxy (header `x-voms_user).
If neither is present, a `-` is printed.

Only the Subject DN is quoted because it can include spaces.