Add support for an "in progress" endpoint
The endpoint can be used by GEMSS to query storm-tape and know what its
view of the world is.
The implementation relies on computing the status of all the stage
requests not yet completed and extracting the files whose recall is still
in progress. The list of files is then included in the response as plain
text. At the moment there is no limit on the number of files to return.

See issue #14