Set the CC env var to point to a recent gcc
The CC var is used by CMake to detect the C compiler, which is then
used to verify for example the presence of a threads library.
Without that var, CMake finds the native and old version of gcc,
which doesn't accept the -fsanitize=undefined option, failing the
check.
Do the same for the C++ compiler, although not strictly needed, since
currently there is no native C++ compiler installed.