Add support to configure mirror mode
If mirror mode is enabled, some read and especially write checks on
the root directory of the storage areas are skipped, because storm-tape
in mirror mode will not need to write on the filesystem so not to
intefere with another running storm-tape.

The name "mirror mode" is temporary, just to connect it to the NGINX
mirror module.