Add an InProgressRequest
The request contains two optional parameters, passed as a URL query string:
- n, the max number of files to retrieve (default 1'000)
- precise, the level of precision of the answer (default 0)

If precise is 0, the files in progress are simply taken from the db. If
precise is 1, the files in progress are taken from the db and then their
current status is checked on the filesystem; if their status has changed,
the db is *not* updated, since the operation is costly (at the moment).