Split locality operation into more primitive ones
The computation of a file locality involves multiple system calls, whose
partial results are not made available to the caller, which is then
sometimes forced to perform them multiple times. This change removes the
locality operation from the Storage interface, replacing it with three
more primitive ones:
- to check if a file is being recalled
- to get file size information about a file (size and if stub)
- to check if a file is on tape
All three return also a possible error code if the underlying call
fails.